[HAL-757] Server-group-scoped-roles not respected in Domain mode
Server-group modify buttons now appear for all users who are part of a valid server-group. If the user does not have the appropriate write permissions for a specific deployement/group, then the request will fail and a console message will be displayed.

Request processing panel is now hidded when a deployement cannot be removed from the Content Repository.